pinfsc50-package {pinfsc50}	R Documentation

pinfsc50: Sequence ('FASTA'), Annotation ('GFF') and Variants ('VCF') for 17 Samples of 'P. Infestans" and 1 'P. Mirabilis'

Description

Genomic data for the plant pathogen "Phytophthora infestans." It includes a variant file ('VCF'), a sequence file ('FASTA') and an annotation file ('GFF'). This package is intended to be used as example data for packages that work with genomic data.

Author(s)

Maintainer: Brian J. Knaus briank.lists@gmail.com

Authors:

Niklaus J. Grunwald grunwaln@science.oregonstate.edu

[Package pinfsc50 version 1.3.0 Index]